// src/app/models/setting.interface.ts
export interface ISetting {
    isMobileView: boolean;
    isTabletView: boolean;
    isMobileMenuOpen: boolean;
    isLightTheme: boolean;
    isSettingsOpen: boolean;
    isMenuTop: boolean;
    isMenuCollapsed: boolean;
    isBorderless: boolean;
    isSquaredBorders: boolean;
    isFixedWidth: boolean;
    isMenuShadow: boolean;
}